Widen topic_id to uint32_t, rename comcore to rlcore, add named topics
topic_id grows from uint16_t to uint32_t across the wire format
(RelinkHeader, register/beacon discovery packets) in both C++ and
Python, verified byte-for-byte cross-language compatible.

Renames the comcore discovery mode to rlcore throughout, including the
standalone daemon (com-core/ -> rlcore/, relink_com_core.* ->
relink_rlcore.*), the client-side API (set_com_core -> set_rlcore,
ComCoreConfig -> RlCoreConfig, etc.), and all references in both
languages.

Adds named topics: advertise/subscribe/publish and the _image variants
now accept a human-readable string, hashed (FNV-1a) to the uint32_t
topic_id that goes on the wire, so both sides just need to agree on a
name instead of a hand-assigned number. Each node keeps a local name
registry (topic_name_for()/rltopic_list()) and throws on a genuine hash
collision or on landing on the reserved NAT-punch id. Covered by new
test_topic_hash test suites in both languages, including a real
brute-force-found collision fixture shared between them.

Updates hello_relink examples and the README to demonstrate the string
topic API alongside the existing numeric one.

+++ relink/include/relink/beacon.hpp
@@ -5,7 +5,7 @@
 //
 // Wire layout (little-endian, matches wire.hpp's BeaconPacketHeader):
 //   BeaconPacket = BeaconPacketHeader(8B: node_ip, node_port, topic_count)
-//                  + topic_count * uint16_t topic_ids
+//                  + topic_count * uint32_t topic_ids
 
 #pragma once
 
@@ -23,11 +23,11 @@ enum class BeaconDecodeResult { Ok, TooShort, LengthMismatch };
 
 inline BeaconEncodeResult encode_beacon_packet(
     uint32_t node_ip, uint16_t node_port,
-    const uint16_t* topic_ids, uint16_t topic_count,
+    const uint32_t* topic_ids, uint16_t topic_count,
     uint8_t* out, size_t out_capacity, size_t* out_len) {
     if (topic_count > kMaxBeaconTopics) return BeaconEncodeResult::TooManyTopics;
 
-    const size_t needed = sizeof(BeaconPacketHeader) + size_t(topic_count) * sizeof(uint16_t);
+    const size_t needed = sizeof(BeaconPacketHeader) + size_t(topic_count) * sizeof(uint32_t);
     if (out_capacity < needed) return BeaconEncodeResult::BufferTooSmall;
 
     BeaconPacketHeader hdr{};
@@ -39,7 +39,7 @@ inline BeaconEncodeResult encode_beacon_packet(
     std::memcpy(out + off, &hdr, sizeof(hdr));
     off += sizeof(hdr);
     for (uint16_t i = 0; i < topic_count; ++i) {
-        uint16_t t = topic_ids[i];
+        uint32_t t = topic_ids[i];
         std::memcpy(out + off, &t, sizeof(t));
         off += sizeof(t);
     }
@@ -58,7 +58,7 @@ inline BeaconDecodeResult decode_beacon_packet(const uint8_t* buf, size_t len, D
     if (len < sizeof(BeaconPacketHeader)) return BeaconDecodeResult::TooShort;
     BeaconPacketHeader hdr{};
     std::memcpy(&hdr, buf, sizeof(hdr));
-    const size_t needed = sizeof(hdr) + size_t(hdr.topic_count) * sizeof(uint16_t);
+    const size_t needed = sizeof(hdr) + size_t(hdr.topic_count) * sizeof(uint32_t);
     if (len != needed) return BeaconDecodeResult::LengthMismatch;
 
     out->node_ip = hdr.node_ip;
@@ -68,9 +68,9 @@ inline BeaconDecodeResult decode_beacon_packet(const uint8_t* buf, size_t len, D
     return BeaconDecodeResult::Ok;
 }
 
-inline uint16_t beacon_topic_at(const DecodedBeacon& b, size_t i) {
-    uint16_t v;
-    std::memcpy(&v, b.topic_ids_raw + i * sizeof(uint16_t), sizeof(v));
+inline uint32_t beacon_topic_at(const DecodedBeacon& b, size_t i) {
+    uint32_t v;
+    std::memcpy(&v, b.topic_ids_raw + i * sizeof(uint32_t), sizeof(v));
     return v;
 }
